\docType{data}
\name{datenames}
\alias{datenames}
\title{A typology of publication history terms}
\format{list (stored as YAML file)}
\usage{
datenames
}
\description{
A list mapping various publisher descriptions of
publication dates to standardized names.  See
\code{get_pub_history} for a description of names
}
\keyword{datasets}

