\name{get_dois}
\alias{get_dois}
\title{Use crossref metadata search to get all DOIs for a journal over a time period}
\usage{
get_dois(ISSN, years)
}
\arguments{
  \item{ISSN}{The journal ISSN number}

  \item{years}{a vector of years for which}
}
\description{
Use crossref metadata search to get all DOIs for a journal
over a time period
}

